Fix automigrate
