+++ lib/cloudant.js
@@ -706,13 +706,15 @@ Cloudant.prototype.selectModel = function(model, migrate) {
     }
   }
 
+  var idName = this.idName(model);
   debug('Cloudant.prototype.selectModel use %j', dbName);
   this.pool[model] = {
     mo: mo,
+    dateFields: dateFields,
     db: this.cloudant.use(dbName),
+    idName: idName,
     modelView: modelView,
     modelSelector: modelSelector,
-    dateFields: dateFields,
   };
 
   return this.pool[model];
@@ -1125,8 +1127,9 @@ Cloudant.prototype._findRecursive = function(mo, query, docs, include, cb) {
     if (err) return cb(err);
 
     // only sort numeric id if the id type is of Number
-    var idName = self.idName(mo.mo.model.modelName);
-    if (mo.mo.properties[idName].type.name === 'Number' && query.sort)
+    var idName = self.getIdName(mo.mo.model.modelName);
+    if (!!idName && mo.mo.properties[idName].type.name === 'Number'
+      && query.sort)
       self._sortNumericId(rst.docs, query.sort);
 
     // work around for issue
@@ -1192,6 +1195,32 @@ Cloudant.prototype._sortNumericId = function(docs, filter) {
     }
   });
 };
+
+/**
+ * Return idName for model existing in this.pool
+ * 
+ * Apply to the following scenario:
+ * ```javascript
+ * var Test;
+ * Test = db.define('Test', {oldid: String});
+ * db.automigrate('Test', function(err) {
+ *   Test = db.define('Test', {newid: String});
+ *   db.automigrate('Test', cb);  
+ * });
+ * ```
+ * 
+ * `automigrate` first destroy all old data then autoupdate, which
+ * also updates `this.pool` with the new model config, but it still
+ * needs the old id name when destroy the data.
+ * 
+ * @param {String} model The model name
+ */
+Cloudant.prototype.getIdName = function(model) {
+  var self = this;
+  var cachedModel = self.pool[model];
+  if (!!cachedModel) return cachedModel.idName;
+  else return self.idName(model);
+}
 
 /**
  * Check if a type is string
