+++ package.json
@@ -33,7 +33,7 @@
     "dockerode": "^2.4.3",
     "eslint": "^2.13.1",
     "eslint-config-loopback": "^4.0.0",
-    "loopback-datasource-juggler": "^3.0.0",
+    "loopback-datasource-juggler": "git://github.com/strongloop/loopback-datasource-juggler.git#couchdb/test",
     "mocha": "^2.3.4",
     "ms": "^2.0.0",
     "rc": "^1.1.5",
