+++ test/regexp.test.js
@@ -21,9 +21,7 @@ describe('cloudant regexp', function() {
     Foo = db.define('Foo', {
       bar: {type: String, index: true},
     });
-    db.once('connected', function() {
-      db.automigrate(done);
-    });
+    db.automigrate(done);
   });
   it('create some foo', function(done) {
     var foos = Array.apply(null, {length: N}).map(function(n, i) {
