+++ test/update.test.js
@@ -33,9 +33,7 @@ describe('updateOrCreate', function() {
       price: {type: Number},
     }, {forceId: false});
 
-    db.once('connected', function() {
-      db.automigrate(done);
-    });
+    db.automigrate(done);
   });
 
   it('creates when model instance does not exist', function(done) {
@@ -127,10 +125,7 @@ describe('updateAll', function() {
       description: {type: String},
       price: {type: Number},
     }, {forceId: false});
-
-    db.once('connected', function() {
-      db.automigrate(done);
-    });
+    db.automigrate(done);
   });
 
   beforeEach(function(done) {
@@ -248,11 +243,8 @@ describe('bulkReplace', function() {
       description: {type: String},
       price: {type: Number},
     }, {forceId: false});
-
-    db.once('connected', function() {
-      db.automigrate(function(err) {
-        Product.create(breads, done);
-      });
+    db.automigrate(function(err) {
+      Product.create(breads, done);
     });
   });
 
@@ -310,10 +302,8 @@ describe('updateAttributes', function() {
       price: {type: Number},
     }, {forceId: false, updateOnLoad: true});
 
-    db.once('connected', function() {
-      db.automigrate(function(err) {
-        Product.create(bread, done);
-      });
+    db.automigrate(function(err) {
+      Product.create(bread, done);
     });
   });
